Fixed the "days as year" UX bug when upgrading a pad
Seems like the KSPUtil function wasn't working as intended, so I wrote my own and put it in KCT_Utilities. My first work in C#, so please review it. I haven't compiled KCT and tested it (not sure how, new to modding), but the function was tested in a .net fiddle so I'm fairly sure it'll work.